+++ backend/.php-cs-fixer.dist.php
@@ -1,0 +1,19 @@
+<?php
+
+$finder = PhpCsFixer\Finder::create()
+    ->in(__DIR__)
+    ->exclude('bootstrap')
+    ->exclude('storage')
+    ->exclude('vendor')
+    ->notPath('src/public')
+    ->notPath('src/readme.md')
+    ->notPath('src/server.php');
+
+return (new PhpCsFixer\Config())
+    ->setRules([
+        '@PSR12' => true,
+        'array_syntax' => ['syntax' => 'short'],
+        'ordered_imports' => ['sort_algorithm' => 'alpha'],
+        'no_unused_imports' => true,
+    ])
+    ->setFinder($finder);